+++ app/server/src/main/scala/org/bitcoins/server/Main.scala
@@ -21,9 +21,12 @@ import org.bitcoins.wallet.models.AccountDAO
 import org.bitcoins.wallet.{LockedWallet, Wallet}
 
 import scala.concurrent.duration._
-import scala.concurrent.{Await, Future}
+import scala.concurrent.{Await, ExecutionContext, Future}
 
 object Main extends App {
+  implicit val system = ActorSystem("bitcoin-s")
+  implicit val ec: ExecutionContext = system.dispatcher
+
   implicit val conf = {
     val dataDirIndexOpt = args.zipWithIndex
       .find(_._1.toLowerCase == "--datadir")
@@ -45,8 +48,7 @@ object Main extends App {
           "Either Neutrino or SPV mode should be enabled")
   implicit val chainConf: ChainAppConfig = conf.chainConf
 
-  implicit val system = ActorSystem("bitcoin-s")
-  import system.dispatcher
+
 
   val peerSocket =
     parseInetSocketAddress(nodeConf.peers.head, nodeConf.network.port)
