+++ chain/src/main/scala/org/bitcoins/chain/config/ChainAppConfig.scala
@@ -16,16 +16,20 @@ import scala.concurrent.{ExecutionContext, Future}
   */
 case class ChainAppConfig(
     private val directory: Path,
-    private val confs: Config*)
-    extends AppConfig {
+    private val confs: Config*)(override implicit val ec: ExecutionContext)
+    extends AppConfig
+    with ChainDbManagement
+    with JdbcProfileComponent {
+
   override protected[bitcoins] def configOverrides: List[Config] = confs.toList
-  override protected[bitcoins] val moduleName: String = "chain"
+  override protected[bitcoins] def moduleName: String = "chain"
   override protected[bitcoins] type ConfigType = ChainAppConfig
   override protected[bitcoins] def newConfigOfType(
       configs: Seq[Config]): ChainAppConfig =
     ChainAppConfig(directory, configs: _*)
-
   protected[bitcoins] def baseDatadir: Path = directory
+
+  override def appConfig: ChainAppConfig = this
 
   /**
     * Checks whether or not the chain project is initialized by
@@ -33,8 +37,7 @@ case class ChainAppConfig(
     * header table
     */
   def isInitialized()(implicit ec: ExecutionContext): Future[Boolean] = {
-    val bhDAO =
-      BlockHeaderDAO()(ec = implicitly[ExecutionContext], appConfig = this)
+    val bhDAO = BlockHeaderDAO()(ec,appConfig)
     val isDefinedOptF = {
       bhDAO.read(chain.genesisBlock.blockHeader.hashBE).map(_.isDefined)
     }
@@ -53,7 +56,7 @@ case class ChainAppConfig(
     * and inserts preliminary data like the genesis block header
     * */
   override def initialize()(implicit ec: ExecutionContext): Future[Unit] = {
-    val numMigrations = ChainDbManagement.migrate(this)
+    val numMigrations = migrate()
 
     logger.info(s"Applied ${numMigrations} to chain project")
 
@@ -66,8 +69,7 @@ case class ChainAppConfig(
           BlockHeaderDbHelper.fromBlockHeader(height = 0,
                                               bh =
                                                 chain.genesisBlock.blockHeader)
-        val blockHeaderDAO =
-          BlockHeaderDAO()(ec = implicitly[ExecutionContext], appConfig = this)
+        val blockHeaderDAO = BlockHeaderDAO()(ec,appConfig)
         val bhCreatedF = blockHeaderDAO.create(genesisHeader)
         bhCreatedF.flatMap { _ =>
           logger.info(s"Inserted genesis block header into DB")
@@ -97,6 +99,6 @@ object ChainAppConfig {
   /** Constructs a chain verification configuration from the default Bitcoin-S
     * data directory and given list of configuration overrides.
     */
-  def fromDefaultDatadir(confs: Config*): ChainAppConfig =
+  def fromDefaultDatadir(confs: Config*)(implicit ec: ExecutionContext): ChainAppConfig =
     ChainAppConfig(AppConfig.DEFAULT_BITCOIN_S_DATADIR, confs: _*)
 }
