+++ chain/src/main/scala/org/bitcoins/chain/models/CompactFilterHeaderTable.scala
@@ -2,8 +2,6 @@ package org.bitcoins.chain.models
 
 import org.bitcoins.core.crypto.DoubleSha256DigestBE
 import org.bitcoins.core.gcs.FilterHeader
-import slick.lifted.Tag
-import slick.jdbc.SQLiteProfile.api._
 
 case class CompactFilterHeaderDb(
     hashBE: DoubleSha256DigestBE,
@@ -33,28 +31,4 @@ object CompactFilterHeaderDbHelper {
       blockHashBE = blockHash,
       height = height
     )
-}
-
-class CompactFilterHeaderTable(tag: Tag)
-    extends Table[CompactFilterHeaderDb](tag, "cfheaders") {
-  import org.bitcoins.db.DbCommonsColumnMappers._
-
-  def hash = column[DoubleSha256DigestBE]("hash", O.PrimaryKey)
-
-  def filterHash = column[DoubleSha256DigestBE]("filter_hash")
-
-  def previousFilterHeader =
-    column[DoubleSha256DigestBE]("previous_filter_header")
-
-  def blockHash = column[DoubleSha256DigestBE]("block_hash")
-
-  def height = column[Int]("height")
-
-  def heightIndex = index("cfheaders_height_index", height)
-
-  def blockHashIndex = index("cfheaders_block_hash_index", blockHash)
-
-  override def * = {
-    (hash, filterHash, previousFilterHeader, blockHash, height) <> (CompactFilterHeaderDb.tupled, CompactFilterHeaderDb.unapply)
-  }
 }
