+++ chain/src/main/scala/org/bitcoins/chain/models/CompactFilterTable.scala
@@ -4,8 +4,6 @@ import org.bitcoins.core.crypto.DoubleSha256DigestBE
 import org.bitcoins.core.gcs.{BlockFilter, FilterType, GolombFilter}
 import org.bitcoins.core.util.CryptoUtil
 import scodec.bits.ByteVector
-import slick.lifted.Tag
-import slick.jdbc.SQLiteProfile.api._
 
 case class CompactFilterDb(
     hashBE: DoubleSha256DigestBE,
@@ -43,27 +41,4 @@ object CompactFilterDbHelper {
                     filterBytes,
                     height,
                     blockHash)
-}
-
-class CompactFilterTable(tag: Tag)
-    extends Table[CompactFilterDb](tag, "cfilters") {
-  import org.bitcoins.db.DbCommonsColumnMappers._
-
-  def hash = column[DoubleSha256DigestBE]("hash")
-
-  def filterType = column[FilterType]("filter_type")
-
-  def bytes = column[ByteVector]("bytes")
-
-  def height = column[Int]("height")
-
-  def blockHash = column[DoubleSha256DigestBE]("block_hash", O.PrimaryKey)
-
-  def heightIndex = index("cfilters_height_index", height)
-
-  def hashIndex = index("cfilters_hash_index", hash)
-
-  override def * = {
-    (hash, filterType, bytes, height, blockHash) <> (CompactFilterDb.tupled, CompactFilterDb.unapply)
-  }
 }
