+++ db-commons-test/src/test/scala/org/bitcoins/db/DbManagementTest.scala
@@ -7,33 +7,57 @@ import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.node.db.NodeDbManagement
 import org.bitcoins.testkit.BitcoinSTestAppConfig
 import org.bitcoins.testkit.BitcoinSTestAppConfig.ProjectType
-import org.bitcoins.testkit.util.BitcoinSUnitTest
+import org.bitcoins.testkit.util.{BitcoinSAsyncTest, BitcoinSUnitTest}
 import org.bitcoins.wallet.config.WalletAppConfig
 import org.bitcoins.wallet.db.WalletDbManagement
 
-class DbManagementTest extends BitcoinSUnitTest {
+import scala.concurrent.ExecutionContext
+
+class DbManagementTest extends BitcoinSAsyncTest {
 
   def dbConfig(project: ProjectType): Config = {
     BitcoinSTestAppConfig.configWithMemoryDb(Some(project))
   }
+
+  def createChainDbManagement(chainAppConfig: ChainAppConfig): ChainDbManagement = new ChainDbManagement with JdbcProfileComponent {
+    override val ec: ExecutionContext = system.dispatcher
+
+    override def appConfig: ChainAppConfig = chainAppConfig
+  }
+
+  def createWalletDbManagement(walletAppConfig: WalletAppConfig): WalletDbManagement = new WalletDbManagement with JdbcProfileComponent {
+    override val ec: ExecutionContext = system.dispatcher
+
+    override def appConfig: WalletAppConfig = walletAppConfig
+  }
+
+  def createNodeDbManagement(nodeAppConfig: NodeAppConfig): NodeDbManagement = new NodeDbManagement with JdbcProfileComponent {
+    override val ec: ExecutionContext = system.dispatcher
+
+    override def appConfig: NodeAppConfig = nodeAppConfig
+  }
+
   it must "run migrations for chain db" in {
     val chainAppConfig = ChainAppConfig(BitcoinSTestAppConfig.tmpDir(),
                                         dbConfig(ProjectType.Chain))
-    val result = ChainDbManagement.migrate(chainAppConfig)
+    val chainDbManagement = createChainDbManagement(chainAppConfig)
+    val result = chainDbManagement.migrate()
     assert(result == 1)
   }
 
   it must "run migrations for wallet db" in {
     val walletAppConfig = WalletAppConfig(BitcoinSTestAppConfig.tmpDir(),
                                           dbConfig(ProjectType.Wallet))
-    val result = WalletDbManagement.migrate(walletAppConfig)
+    val walletDbManagement = createWalletDbManagement(walletAppConfig)
+    val result = walletDbManagement.migrate()
     assert(result == 4)
   }
 
   it must "run migrations for node db" in {
     val nodeAppConfig =
       NodeAppConfig(BitcoinSTestAppConfig.tmpDir(), dbConfig(ProjectType.Node))
-    val result = NodeDbManagement.migrate(nodeAppConfig)
+    val nodeDbManagement = createNodeDbManagement(nodeAppConfig)
+    val result = nodeDbManagement.migrate()
     assert(result == 1)
   }
 }
