+++ db-commons/src/main/scala/org/bitcoins/db/CRUD.scala
@@ -1,11 +1,90 @@
 package org.bitcoins.db
 
+import java.nio.file.{Files, Path, Paths}
 import java.sql.SQLException
 
-import org.bitcoins.core.config.MainNet
-import slick.jdbc.SQLiteProfile.api._
+import org.bitcoins.core.util.BitcoinSLogger
+import slick.basic.DatabaseConfig
+import slick.dbio.{DBIOAction, NoStream}
+import slick.jdbc.{JdbcProfile}
 
 import scala.concurrent.{ExecutionContext, Future}
+import scala.util.{Failure, Success, Try}
+
+trait JdbcProfileComponent extends BitcoinSLogger {
+
+  def appConfig: AppConfig
+
+  /**
+   * The configuration details for connecting/using the database for our projects
+   * that require datbase connections
+   */
+  val dbConfig: DatabaseConfig[JdbcProfile] = {
+    val slickDbConfig = {
+      Try {
+        DatabaseConfig.forConfig[JdbcProfile](path = appConfig.moduleName,
+          config = appConfig.config)
+      } match {
+        case Success(value) =>
+          value
+        case Failure(exception) =>
+          logger.error(s"Error when loading database from config: $exception")
+          logger.error(s"Configuration: ${appConfig.config.asReadableJson}")
+          throw exception
+      }
+    }
+
+    logger.debug(s"Resolved DB config: ${slickDbConfig.config.asReadableJson}")
+
+    val _ = createDbFileIfDNE()
+
+    slickDbConfig
+  }
+
+  val profile: JdbcProfile = dbConfig.profile
+  import profile.api._
+
+  lazy val jdbcUrl: String = {
+    dbConfig.config.getString("db.url")
+  }
+
+  /** The database we are connecting to */
+  lazy val database: Database = {
+    dbConfig.db
+  }
+
+  /** The path where our DB is located */
+  // todo: what happens to this if we
+  // dont use SQLite?
+  lazy val dbPath: Path = {
+    val pathStr = appConfig.config.getString(s"${appConfig.moduleName}.db.path")
+    val path = Paths.get(pathStr)
+    logger.debug(s"DB path: $path")
+    path
+  }
+
+  /** The name of our database */
+  // todo: what happens to this if we
+  // dont use SQLite?
+  lazy val dbName: String = {
+    appConfig.config.getString(s"${appConfig.moduleName}.db.name")
+  }
+
+  private def createDbFileIfDNE(): Unit = {
+    //should add a check in here that we are using sqlite
+    if (!Files.exists(dbPath)) {
+      val _ = {
+        logger.debug(s"Creating database directory=$dbPath")
+        Files.createDirectories(dbPath)
+        val dbFilePath = dbPath.resolve(dbName)
+        logger.debug(s"Creating database file=$dbFilePath")
+        Files.createFile(dbFilePath)
+      }
+
+      ()
+    }
+  }
+}
 
 /**
   * Created by chris on 9/8/16.
@@ -16,15 +95,16 @@ import scala.concurrent.{ExecutionContext, Future}
   * the table and the database you are connecting to.
   */
 abstract class CRUD[T, PrimaryKeyType](
-    implicit private val config: AppConfig,
-    private val ec: ExecutionContext)
-    extends DatabaseLogger {
+    private implicit val ec: ExecutionContext, override val appConfig: AppConfig)
+    extends JdbcProfileComponent {
+
+  import profile.api._
 
   /** The table inside our database we are inserting into */
-  val table: TableQuery[_ <: Table[T]]
+  val table: profile.api.TableQuery[_ <: profile.api.Table[T]]
 
   /** Binding to the actual database itself, this is what is used to run querys */
-  def database: SafeDatabase = SafeDatabase(config)
+  def safeDatabase: SafeDatabase = SafeDatabase(this)
 
   /**
     * create a record in the database
@@ -33,7 +113,7 @@ abstract class CRUD[T, PrimaryKeyType](
     * @return the inserted record
     */
   def create(t: T): Future[T] = {
-    logger.trace(s"Writing $t to DB with config: ${config.config}")
+    logger.trace(s"Writing $t to DB with config: ${appConfig.config}")
     createAll(Vector(t)).map(_.head)
   }
 
@@ -46,9 +126,9 @@ abstract class CRUD[T, PrimaryKeyType](
     * @return Option[T] - the record if found, else none
     */
   def read(id: PrimaryKeyType): Future[Option[T]] = {
-    logger.trace(s"Reading from DB with config: ${config.config}")
+    logger.trace(s"Reading from DB with config: ${appConfig.config}")
     val query = findByPrimaryKey(id)
-    val rows: Future[Seq[T]] = database.run(query.result)
+    val rows: Future[Seq[T]] = safeDatabase.run(query.result)
     rows.map(_.headOption)
   }
 
@@ -66,10 +146,10 @@ abstract class CRUD[T, PrimaryKeyType](
   def updateAll(ts: Vector[T]): Future[Vector[T]] = {
     val query = findAll(ts)
     val actions = ts.map(t => query.update(t))
-    val affectedRows: Future[Vector[Int]] = database.run(DBIO.sequence(actions))
+    val affectedRows: Future[Vector[Int]] = safeDatabase.run(DBIO.sequence(actions))
     val updatedTs = findAll(ts)
     affectedRows.flatMap { _ =>
-      database.runVec(updatedTs.result)
+      safeDatabase.runVec(updatedTs.result)
     }
   }
 
@@ -82,14 +162,14 @@ abstract class CRUD[T, PrimaryKeyType](
   def delete(t: T): Future[Int] = {
     logger.debug("Deleting record: " + t)
     val query: Query[Table[_], T, Seq] = find(t)
-    database.run(query.delete)
+    safeDatabase.run(query.delete)
   }
 
   /**
     * delete all records from the table
     */
   def deleteAll(): Future[Int] =
-    database.run(table.delete)
+    safeDatabase.run(table.delete)
 
   /**
     * insert the record if it does not exist, update it if it does
@@ -102,9 +182,9 @@ abstract class CRUD[T, PrimaryKeyType](
   /** Upserts all of the given ts in the database, then returns the upserted values */
   def upsertAll(ts: Vector[T]): Future[Vector[T]] = {
     val actions = ts.map(t => table.insertOrUpdate(t))
-    val result: Future[Vector[Int]] = database.run(DBIO.sequence(actions))
+    val result: Future[Vector[Int]] = safeDatabase.run(DBIO.sequence(actions))
     val findQueryFuture = result.map(_ => findAll(ts).result)
-    findQueryFuture.flatMap(database.runVec(_))
+    findQueryFuture.flatMap(safeDatabase.runVec(_))
   }
 
   /**
@@ -118,7 +198,7 @@ abstract class CRUD[T, PrimaryKeyType](
 
   /** Finds the rows that correlate to the given primary keys */
   protected def findByPrimaryKeys(
-      ids: Vector[PrimaryKeyType]): Query[Table[_], T, Seq]
+      ids: Vector[PrimaryKeyType]): Query[Table[T], T, Seq]
 
   /**
     * return the row that corresponds with this record
@@ -132,16 +212,17 @@ abstract class CRUD[T, PrimaryKeyType](
 
   /** Finds all elements in the table */
   def findAll(): Future[Vector[T]] =
-    database.run(table.result).map(_.toVector)
+    safeDatabase.run(table.result).map(_.toVector)
 
   /** Returns number of rows in the table */
-  def count(): Future[Int] = database.run(table.length.result)
+  def count(): Future[Int] = safeDatabase.run(table.length.result)
 }
 
-case class SafeDatabase(config: AppConfig) extends DatabaseLogger {
-  implicit private val conf: AppConfig = config
+case class SafeDatabase(jdbcProfile: JdbcProfileComponent) extends BitcoinSLogger {
 
-  import config.database
+  import jdbcProfile.database
+
+  import jdbcProfile.profile.api.actionBasedSQLInterpolation
 
   /**
     * SQLite does not enable foreign keys by default. This query is
@@ -154,11 +235,9 @@ case class SafeDatabase(config: AppConfig) extends DatabaseLogger {
   private def logAndThrowError(
       action: DBIOAction[_, NoStream, _]): PartialFunction[Throwable, Nothing] = {
     case err: SQLException =>
-      if (config.network != MainNet) {
-        logger.error(
-          s"Error when executing query ${action.getDumpInfo.getNamePlusMainInfo}")
-        logger.error(s"$err")
-      }
+      logger.error(
+        s"Error when executing query ${action.getDumpInfo.getNamePlusMainInfo}")
+      logger.error(s"$err")
       throw err
   }
