+++ db-commons/src/main/scala/org/bitcoins/db/CRUDAutoInc.scala
@@ -1,30 +1,29 @@
 package org.bitcoins.db
 
-import slick.dbio.Effect.Write
-import slick.jdbc.SQLiteProfile.api._
-
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext
 
-abstract class CRUDAutoInc[T <: DbRowAutoInc[T]](
-    implicit config: AppConfig,
-    ec: ExecutionContext)
-    extends CRUD[T, Long] {
+abstract class CRUDAutoInc[T <: DbRowAutoInc[T]](implicit ec: ExecutionContext, override val appConfig: AppConfig)
+    extends CRUD[T, Long]()(ec, appConfig) with TableAutoIncComponent[T] {
+  import profile.api._
 
   /** The table inside our database we are inserting into */
-  override val table: TableQuery[_ <: TableAutoInc[T]]
+  override val table: profile.api.TableQuery[_ <: TableAutoInc[T]]
 
   override def createAll(ts: Vector[T]): Future[Vector[T]] = {
-    val query = table
-      .returning(table.map(_.id))
-      .into((t, id) => t.copyWithId(id = id))
-    val actions: DBIOAction[query.MultiInsertResult, NoStream, Write] =
-      query.++=(ts)
-    database.runVec(actions)
+    val idQuery = table.map(_.id)
+    val idAutoInc = table.returning(idQuery)
+    val query = {
+      idAutoInc.into((t, id) => t.copyWithId(id = id))
+    }
+    val actions = query.++=(ts)
+    safeDatabase.runVec(actions)
   }
 
-  override def findByPrimaryKeys(ids: Vector[Long]): Query[Table[_], T, Seq] = {
-    table.filter(_.id.inSet(ids))
+  override def findByPrimaryKeys(ids: Vector[Long]): Query[TableAutoInc[T], T, Seq] = {
+    table.filter { t =>
+      t.id.inSet(ids)
+    }
   }
 
   override def findAll(ts: Vector[T]): Query[Table[_], T, Seq] = {
@@ -32,3 +31,22 @@ abstract class CRUDAutoInc[T <: DbRowAutoInc[T]](
     findByPrimaryKeys(ids)
   }
 }
+
+
+/** Defines a table that has an auto incremented fields that is named id.
+ * This is useful for things we want to store that don't have an
+ * inherent id such as a hash.
+ * @param tag
+ * @param tableName
+ * @tparam T
+ */
+trait TableAutoIncComponent[T <: DbRowAutoInc[T]] { self: CRUDAutoInc[T] =>
+  import profile.api._
+
+  abstract class TableAutoInc[T](tag: profile.api.Tag, tableName: String) extends
+    profile.api.Table[T](tag,tableName) {
+    def id: Rep[Long] = column[Long]("id", O.PrimaryKey, O.AutoInc)
+  }
+}
+
+
