+++ db-commons/src/main/scala/org/bitcoins/db/DbManagement.scala
@@ -1,33 +1,28 @@
 package org.bitcoins.db
 
+import org.bitcoins.core.util.BitcoinSLogger
 import org.flywaydb.core.Flyway
 import org.flywaydb.core.api.FlywayException
-import slick.jdbc.SQLiteProfile.api._
 
 import scala.concurrent.{ExecutionContext, Future}
 
-abstract class DbManagement extends DatabaseLogger {
-  def allTables: List[TableQuery[_ <: Table[_]]]
+trait DbManagement extends BitcoinSLogger { _: JdbcProfileComponent =>
+  import profile.api._
+
+  def allTables: List[TableQuery[Table[_]]]
 
   /** Lists all tables in the given database */
-  def listTables(db: Database): Future[Vector[SQLiteTableInfo]] = {
+  def listTables: Future[Vector[SQLiteTableInfo]] = {
     import DbCommonsColumnMappers._
     val query = sql"SELECT * FROM sqlite_master where type='table'"
       .as[SQLiteTableInfo]
-    db.run(query)
+    database.run(query)
   }
 
-  /** Lists all tables in the given database */
-  def listTables(db: SafeDatabase): Future[Vector[SQLiteTableInfo]] =
-    listTables(db.config.database)
-
   /** Creates all tables in our table list, in one SQL transaction */
-  def createAll()(
-      implicit config: AppConfig,
-      ec: ExecutionContext
-  ): Future[Unit] = {
-    val tables = allTables.map(_.baseTableRow.tableName).mkString(", ")
-    logger.debug(s"Creating tables: $tables")
+  def createAll()(implicit ec: ExecutionContext): Future[Unit] = {
+/*    val tables = allTables.map(_.baseTableRow.tableName).mkString(", ")
+    logger.debug(s"Creating tables: $tables")*/
 
     val query = {
       val querySeq =
@@ -42,13 +37,10 @@ abstract class DbManagement extends DatabaseLogger {
       DBIO.seq(querySeq: _*).transactionally
     }
 
-    import config.database
-    database.run(query).map(_ => logger.debug(s"Created tables: $tables"))
+    database.run(query).map(_ => logger.debug(s"Created tables"))
   }
 
-  def dropAll()(
-      implicit config: AppConfig,
-      ec: ExecutionContext): Future[Unit] = {
+  def dropAll()(implicit ec: ExecutionContext): Future[Unit] = {
     Future.sequence(allTables.reverse.map(dropTable(_))).map(_ => ())
   }
 
@@ -66,22 +58,18 @@ abstract class DbManagement extends DatabaseLogger {
   /** Creates the given table */
   def createTable(
       table: TableQuery[_ <: Table[_]],
-      createIfNotExists: Boolean = true)(
-      implicit config: AppConfig,
-      ec: ExecutionContext): Future[Unit] = {
+      createIfNotExists: Boolean = true)(implicit ec: ExecutionContext): Future[Unit] = {
     val tableName = table.baseTableRow.tableName
     logger.debug(
-      s"Creating table $tableName with DB config: ${config.dbConfig.config} ")
+      s"Creating table $tableName with DB config: ${appConfig.config} ")
 
-    import config.database
     val query = createTableQuery(table, createIfNotExists)
     database.run(query).map(_ => logger.debug(s"Created table $tableName"))
   }
 
   def dropTable(
-      table: TableQuery[_ <: Table[_]]
-  )(implicit config: AppConfig): Future[Unit] = {
-    import config.database
+      table: TableQuery[Table[_]]
+  ): Future[Unit] = {
     val result = database.run(table.schema.dropIfExists)
     result
   }
@@ -89,21 +77,21 @@ abstract class DbManagement extends DatabaseLogger {
   /** Executes migrations related to this database
     *
     * @see [[https://flywaydb.org/documentation/api/#programmatic-configuration-java]] */
-  def migrate(appConfig: AppConfig): Int = {
-    val url = appConfig.jdbcUrl
+  def migrate(): Int = {
+    val url = jdbcUrl
     val username = ""
     val password = ""
     //appConfig.dbName is for the format 'walletdb.sqlite' or 'nodedb.sqlite' etc
     //we need to remove the '.sqlite' suffix
-    val dbName = appConfig.dbName.split('.').head.mkString
-    val config = Flyway.configure().locations(s"classpath:${dbName}/migration/")
+    val name = dbName.split('.').head.mkString
+    val config = Flyway.configure().locations(s"classpath:${name}/migration/")
     val flyway = config.dataSource(url, username, password).load
 
     try {
       flyway.migrate()
     } catch {
       case err: FlywayException =>
-        logger(appConfig).warn(
+        logger.warn(
           s"Failed to apply first round of migrations, attempting baseline and re-apply",
           err)
         //maybe we have an existing database, so attempt to baseline the existing
