+++ node/src/main/scala/org/bitcoins/node/config/NodeAppConfig.scala
@@ -4,7 +4,7 @@ import java.nio.file.Path
 
 import com.typesafe.config.Config
 import org.bitcoins.core.util.FutureUtil
-import org.bitcoins.db.AppConfig
+import org.bitcoins.db.{AppConfig, JdbcProfileComponent}
 import org.bitcoins.node.db.NodeDbManagement
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -15,10 +15,12 @@ import scala.concurrent.{ExecutionContext, Future}
   */
 case class NodeAppConfig(
     private val directory: Path,
-    private val confs: Config*)
-    extends AppConfig {
+    private val confs: Config*)(implicit override val ec: ExecutionContext)
+    extends AppConfig
+      with NodeDbManagement
+      with JdbcProfileComponent{
   override protected[bitcoins] def configOverrides: List[Config] = confs.toList
-  override protected[bitcoins] val moduleName: String = "node"
+  override protected[bitcoins] def moduleName: String = "node"
   override protected[bitcoins] type ConfigType = NodeAppConfig
   override protected[bitcoins] def newConfigOfType(
       configs: Seq[Config]): NodeAppConfig =
@@ -26,13 +28,15 @@ case class NodeAppConfig(
 
   protected[bitcoins] def baseDatadir: Path = directory
 
+  override def appConfig: NodeAppConfig = this
+
   /**
     * Ensures correct tables and other required information is in
     * place for our node.
     */
   override def initialize()(implicit ec: ExecutionContext): Future[Unit] = {
     logger.debug(s"Initializing node setup")
-    val numMigrations = NodeDbManagement.migrate(this)
+    val numMigrations = migrate()
 
     logger.info(s"Applied $numMigrations migrations fro the node project")
 
@@ -70,7 +74,7 @@ object NodeAppConfig {
   /** Constructs a node configuration from the default Bitcoin-S
     * data directory and given list of configuration overrides.
     */
-  def fromDefaultDatadir(confs: Config*): NodeAppConfig =
+  def fromDefaultDatadir(confs: Config*)(implicit ec: ExecutionContext): NodeAppConfig =
     NodeAppConfig(AppConfig.DEFAULT_BITCOIN_S_DATADIR, confs: _*)
 
 }
