+++ node/src/main/scala/org/bitcoins/node/db/NodeDbManagement.scala
@@ -1,12 +1,22 @@
 package org.bitcoins.node.db
 
-import org.bitcoins.db.DbManagement
-import org.bitcoins.node.models.BroadcastAbleTransactionTable
-import slick.lifted.TableQuery
+import org.bitcoins.db.{DbManagement, JdbcProfileComponent}
+import org.bitcoins.node.config.NodeAppConfig
+import org.bitcoins.node.models.{BroadcastAbleTransactionDAO}
 
-object NodeDbManagement extends DbManagement {
+import scala.concurrent.ExecutionContext
 
-  private val txTable = TableQuery[BroadcastAbleTransactionTable]
+trait NodeDbManagement extends DbManagement { _: JdbcProfileComponent =>
+
+  import profile.api._
+
+  def ec: ExecutionContext
+
+  override def appConfig: NodeAppConfig
+
+  private lazy val txTable: TableQuery[Table[_]] = {
+    BroadcastAbleTransactionDAO()(appConfig.asInstanceOf[NodeAppConfig],ec).table.asInstanceOf[TableQuery[Table[_]]]
+  }
 
   override val allTables = List(txTable)
