+++ testkit/src/main/scala/org/bitcoins/testkit/fixtures/NodeDAOFixture.scala
@@ -1,17 +1,15 @@
 package org.bitcoins.testkit.fixtures
 
-import org.bitcoins.node.db.NodeDbManagement
 import org.bitcoins.node.models.BroadcastAbleTransactionDAO
 import org.bitcoins.testkit.node.NodeUnitTest
 import org.scalatest._
-import slick.jdbc.SQLiteProfile
 
 case class NodeDAOs(txDAO: BroadcastAbleTransactionDAO)
 
 /** Provides a fixture where all DAOs used by the node projects are provided */
 trait NodeDAOFixture extends NodeUnitTest {
   private lazy val daos = {
-    val tx = BroadcastAbleTransactionDAO(SQLiteProfile)
+    val tx = BroadcastAbleTransactionDAO()
     NodeDAOs(tx)
   }
 
@@ -19,10 +17,10 @@ trait NodeDAOFixture extends NodeUnitTest {
 
   def withFixture(test: OneArgAsyncTest): FutureOutcome =
     makeFixture(build = () =>
-                  NodeDbManagement
-                    .createAll()(nodeConfig, executionContext)
+                  nodeConfig
+                    .createAll()(executionContext)
                     .map(_ => daos),
                 destroy = () =>
-                  NodeDbManagement
-                    .dropAll()(nodeConfig, executionContext))(test)
+                  nodeConfig
+                    .dropAll()(executionContext))(test)
 }
