+++ testkit/src/test/scala/org/bitcoins/testkit/db/AppConfigTest.scala
@@ -1,37 +1,21 @@
 package org.bitcoins.testkit.db
 
-import org.bitcoins.testkit.util.BitcoinSUnitTest
-import org.bitcoins.testkit.Implicits._
-import org.bitcoins.server.BitcoinSAppConfig
-import org.bitcoins.server.BitcoinSAppConfig._
-import com.typesafe.config.ConfigFactory
-import org.bitcoins.core.config.TestNet3
-import org.bitcoins.chain.models.BlockHeaderDAO
-import akka.actor.ActorSystem
-import scala.concurrent.ExecutionContext
-import org.bitcoins.wallet.models.AccountDAO
-import org.bitcoins.testkit.chain.ChainTestUtil
-import org.bitcoins.chain.models.BlockHeaderDb
-import org.bitcoins.chain.models.BlockHeaderDbHelper
-import org.bitcoins.wallet.models.AccountDb
-import org.bitcoins.core.hd.HDAccount
-import org.bitcoins.core.hd.HDCoin
-import org.bitcoins.core.hd.HDPurposes
-import org.bitcoins.core.hd.HDCoinType
-import org.bitcoins.testkit.core.gen.CryptoGenerators
-import org.bitcoins.node.db.NodeDbManagement
-import org.bitcoins.db.DbManagement
-import org.bitcoins.wallet.db.WalletDbManagement
-import org.bitcoins.db.SQLiteTableInfo
-import slick.jdbc.SQLiteProfile.api._
-import org.bitcoins.db.CRUD
 import java.nio.file.Files
+
+import com.typesafe.config.ConfigFactory
+import org.bitcoins.chain.models.BlockHeaderDAO
+import org.bitcoins.core.config.TestNet3
+import org.bitcoins.core.hd.{HDAccount, HDCoin, HDCoinType, HDPurposes}
+import org.bitcoins.db.{CRUD, SQLiteTableInfo}
+import org.bitcoins.server.BitcoinSAppConfig._
 import org.bitcoins.testkit.BitcoinSTestAppConfig
+import org.bitcoins.testkit.Implicits._
+import org.bitcoins.testkit.chain.ChainTestUtil
+import org.bitcoins.testkit.core.gen.CryptoGenerators
+import org.bitcoins.testkit.util.BitcoinSAsyncTest
+import org.bitcoins.wallet.models.{AccountDAO, AccountDb}
 
-class AppConfigTest extends BitcoinSUnitTest {
-
-  val system = ActorSystem()
-  implicit val ec: ExecutionContext = system.dispatcher
+class AppConfigTest extends BitcoinSAsyncTest {
 
   behavior of "BitcoinSAppConfig"
 
@@ -68,60 +52,6 @@ class AppConfigTest extends BitcoinSUnitTest {
     val nodeConf = conf.nodeConf
     assert(chainConf.dbName != walletConf.dbName)
     assert(walletConf.dbName != nodeConf.dbName)
-  }
-
-  it must "be able to write to distinct databases" in {
-    implicit val config = BitcoinSTestAppConfig.getSpvTestConfig()
-    val chainConf = config.chainConf
-    val walletConf = config.walletConf
-    val nodeConf = config.nodeConf
-    val allConfs = List(chainConf, walletConf, nodeConf)
-
-    val bhDAO = BlockHeaderDAO()
-    val accountDAO = AccountDAO()
-
-    allConfs.foreach { conf =>
-      val fullDbPath = conf.dbPath.resolve(conf.dbName)
-      assert(!Files.exists(fullDbPath))
-    }
-
-    val writeF = {
-      for {
-        _ <- config.initialize()
-        _ = {
-          allConfs.foreach { conf =>
-            val fullDbPath = conf.dbPath.resolve(conf.dbName)
-            assert(Files.isRegularFile(fullDbPath))
-          }
-        }
-        _ <- {
-          bhDAO.create(ChainTestUtil.regTestGenesisHeaderDb)
-        }
-        _ <- {
-          val hdAccount =
-            HDAccount(HDCoin(HDPurposes.Legacy, HDCoinType.Bitcoin), 0)
-          val xpub = CryptoGenerators.extPublicKey.sampleSome
-          val account = AccountDb(xpub, hdAccount)
-          accountDAO.create(account)
-        }
-      } yield ()
-    }
-
-    for {
-      _ <- writeF
-      nodeTables <- NodeDbManagement.listTables(bhDAO.database)
-      walletTables <- WalletDbManagement.listTables(accountDAO.database)
-    } yield {
-      def hasTable(tables: Seq[SQLiteTableInfo], dao: CRUD[_, _]): Boolean =
-        tables.exists(_.name == dao.table.baseTableRow.tableName)
-
-      assert(hasTable(walletTables, accountDAO))
-      assert(!hasTable(walletTables, bhDAO))
-
-      assert(hasTable(nodeTables, bhDAO))
-      assert(!hasTable(nodeTables, accountDAO))
-    }
-
   }
 
 }
