+++ wallet-test/src/test/scala/org/bitcoins/wallet/WalletAppConfigTest.scala
@@ -1,6 +1,6 @@
 package org.bitcoins.wallet
 
-import org.bitcoins.testkit.util.BitcoinSUnitTest
+import org.bitcoins.testkit.util.{BitcoinSAsyncTest, BitcoinSUnitTest}
 import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.core.config.TestNet3
 import com.typesafe.config.Config
@@ -9,13 +9,16 @@ import org.bitcoins.core.config.RegTest
 import org.bitcoins.core.config.MainNet
 import org.bitcoins.wallet.config.WalletAppConfig
 import java.nio.file.Paths
+
 import org.bitcoins.core.hd.HDPurposes
 import java.nio.file.Files
+
 import ch.qos.logback.classic.Level
 import java.nio.file.Path
+
 import scala.util.Properties
 
-class WalletAppConfigTest extends BitcoinSUnitTest {
+class WalletAppConfigTest extends BitcoinSAsyncTest {
 
   val tempDir = Files.createTempDirectory("bitcoin-s")
   val config = WalletAppConfig(directory = tempDir)
