+++ wallet/src/main/scala/org/bitcoins/wallet/models/AddressTable.scala
@@ -4,15 +4,8 @@ import org.bitcoins.core.config.NetworkParameters
 import org.bitcoins.core.crypto.{ECPublicKey, Sha256Hash160Digest}
 import org.bitcoins.core.hd._
 import org.bitcoins.core.protocol.script._
-import org.bitcoins.core.protocol.{
-  Bech32Address,
-  BitcoinAddress,
-  P2PKHAddress,
-  P2SHAddress
-}
+import org.bitcoins.core.protocol.{Bech32Address, BitcoinAddress, P2PKHAddress, P2SHAddress}
 import org.bitcoins.core.script.ScriptType
-import slick.jdbc.SQLiteProfile.api._
-import slick.lifted.ProvenShape
 
 sealed trait AddressDb {
   protected type PathType <: HDPath
@@ -133,184 +126,4 @@ object AddressDbHelper {
     case nested: NestedSegWitHDPath => getNestedSegwitAddress(pub, nested, np)
     case segwit: SegWitHDPath       => getSegwitAddress(pub, segwit, np)
   }
-}
-
-/**
-  * todo: this needs design rework.
-  * todo: https://github.com/bitcoin-s/bitcoin-s-core/pull/391#discussion_r274188334
-  */
-class AddressTable(tag: Tag) extends Table[AddressDb](tag, "addresses") {
-  import org.bitcoins.db.DbCommonsColumnMappers._
-
-  def purpose: Rep[HDPurpose] = column("hd_purpose")
-
-  def accountCoin: Rep[HDCoinType] = column("hd_coin")
-
-  def accountIndex: Rep[Int] = column("account_index")
-
-  def accountChainType: Rep[HDChainType] = column("hd_chain_type")
-
-  def addressIndex: Rep[Int] = column("address_index")
-
-  def address: Rep[BitcoinAddress] = column("address", O.PrimaryKey)
-
-  def ecPublicKey: Rep[ECPublicKey] = column("pubkey")
-
-  def hashedPubKey: Rep[Sha256Hash160Digest] = column("hashed_pubkey")
-
-  def scriptType: Rep[ScriptType] = column("script_type")
-
-  def scriptPubKey: Rep[ScriptPubKey] = column("script_pub_key", O.Unique)
-
-  def scriptWitness: Rep[Option[ScriptWitness]] = column("script_witness")
-
-  private type AddressTuple = (
-      HDPurpose,
-      Int,
-      HDCoinType,
-      HDChainType,
-      BitcoinAddress,
-      Option[ScriptWitness],
-      ScriptPubKey,
-      Int,
-      ECPublicKey,
-      Sha256Hash160Digest,
-      ScriptType)
-
-  private val fromTuple: AddressTuple => AddressDb = {
-    case (
-        purpose,
-        accountIndex,
-        accountCoin,
-        accountChain,
-        address,
-        scriptWitnessOpt,
-        scriptPubKey,
-        addressIndex,
-        pubKey,
-        hashedPubKey,
-        scriptType @ _ // what should we do about this? scriptType is inferrable from purpose
-        ) =>
-      (purpose, address, scriptWitnessOpt) match {
-        case (HDPurposes.SegWit,
-              bechAddr: Bech32Address,
-              Some(scriptWitness)) =>
-          val path =
-            SegWitHDPath(coinType = accountCoin,
-                         accountIndex = accountIndex,
-                         chainType = accountChain,
-                         addressIndex = addressIndex)
-
-          SegWitAddressDb(path,
-                          ecPublicKey = pubKey,
-                          hashedPubKey = hashedPubKey,
-                          address = bechAddr,
-                          witnessScript = scriptWitness,
-                          scriptPubKey = scriptPubKey)
-
-        case (HDPurposes.Legacy, legacyAddr: P2PKHAddress, None) =>
-          val path = LegacyHDPath(coinType = accountCoin,
-                                  accountIndex = accountIndex,
-                                  chainType = accountChain,
-                                  addressIndex = addressIndex)
-          LegacyAddressDb(path,
-                          pubKey,
-                          hashedPubKey,
-                          legacyAddr,
-                          scriptPubKey = scriptPubKey)
-
-        case (HDPurposes.NestedSegWit,
-              address: P2SHAddress,
-              Some(scriptWitness)) =>
-          val path = NestedSegWitHDPath(coinType = accountCoin,
-                                        accountIndex = accountIndex,
-                                        chainType = accountChain,
-                                        addressIndex = addressIndex)
-          NestedSegWitAddressDb(path,
-                                pubKey,
-                                hashedPubKey,
-                                address,
-                                witnessScript = scriptWitness,
-                                scriptPubKey = scriptPubKey)
-        case (purpose: HDPurpose, address: BitcoinAddress, scriptWitnessOpt) =>
-          throw new IllegalArgumentException(
-            s"Got invalid combination of HD purpose, address and script witness: $purpose, $address, $scriptWitnessOpt")
-      }
-  }
-
-  private val toTuple: AddressDb => Option[AddressTuple] = {
-    case SegWitAddressDb(path,
-                         pubKey,
-                         hashedPubKey,
-                         address,
-                         scriptWitness,
-                         scriptPubKey) =>
-      Some(
-        (path.purpose,
-         path.account.index,
-         path.coin.coinType,
-         path.chain.chainType,
-         address,
-         Some(scriptWitness),
-         scriptPubKey,
-         path.address.index,
-         pubKey,
-         hashedPubKey,
-         ScriptType.WITNESS_V0_KEYHASH))
-    case LegacyAddressDb(path, pubkey, hashedPub, address, scriptPubKey) =>
-      Some(
-        path.purpose,
-        path.account.index,
-        path.coin.coinType,
-        path.chain.chainType,
-        address,
-        None, // scriptwitness
-        scriptPubKey,
-        path.address.index,
-        pubkey,
-        hashedPub,
-        ScriptType.PUBKEYHASH
-      )
-    case NestedSegWitAddressDb(path,
-                               pubKey,
-                               hashedPubKey,
-                               address,
-                               scriptWitness,
-                               scriptPubKey) =>
-      Some(
-        (path.purpose,
-         path.account.index,
-         path.coin.coinType,
-         path.chain.chainType,
-         address,
-         Some(scriptWitness),
-         scriptPubKey,
-         path.address.index,
-         pubKey,
-         hashedPubKey,
-         ScriptType.SCRIPTHASH))
-  }
-
-  override def * : ProvenShape[AddressDb] =
-    (purpose,
-     accountIndex,
-     accountCoin,
-     accountChainType,
-     address,
-     scriptWitness,
-     scriptPubKey,
-     addressIndex,
-     ecPublicKey,
-     hashedPubKey,
-     scriptType) <> (fromTuple, toTuple)
-
-  val accounts = TableQuery[AccountTable]
-
-  // for some reason adding a type annotation here causes compile error
-  def fk =
-    foreignKey("fk_account",
-               sourceColumns = (purpose, accountCoin, accountIndex),
-               targetTableQuery = accounts) { accountTable =>
-      (accountTable.purpose, accountTable.coinType, accountTable.index)
-    }
 }
