+++ wallet/src/main/scala/org/bitcoins/wallet/models/IncomingTransactionDAO.scala
@@ -1,13 +1,50 @@
 package org.bitcoins.wallet.models
 
+import org.bitcoins.core.crypto.DoubleSha256DigestBE
+import org.bitcoins.core.currency.CurrencyUnit
 import org.bitcoins.wallet.config._
-import slick.lifted.TableQuery
+import slick.lifted.{PrimaryKey, ProvenShape}
 
 import scala.concurrent.ExecutionContext
 
 case class IncomingTransactionDAO()(
     implicit val ec: ExecutionContext,
-    val appConfig: WalletAppConfig)
-    extends TxDAO[IncomingTransactionDb, IncomingTransactionTable] {
+    override val appConfig: WalletAppConfig)
+    extends TxDAO[IncomingTransactionDb] {
+  import profile.api._
   override val table = TableQuery[IncomingTransactionTable]
+
+
+  class IncomingTransactionTable(tag: Tag)
+    extends TxTable[IncomingTransactionDb](tag, "wallet_incoming_txs") {
+
+    import org.bitcoins.db.DbCommonsColumnMappers._
+
+    def txIdBE: Rep[DoubleSha256DigestBE] = column("txIdBE", O.Unique)
+
+    def incomingAmount: Rep[CurrencyUnit] = column("incomingAmount")
+
+    private type IncomingTransactionTuple = (DoubleSha256DigestBE, CurrencyUnit)
+
+    private val fromTuple: IncomingTransactionTuple => IncomingTransactionDb = {
+      case (txId, incomingAmount) =>
+        IncomingTransactionDb(txId, incomingAmount)
+    }
+
+    private val toTuple: IncomingTransactionDb => Option[
+      IncomingTransactionTuple] = tx => Some((tx.txIdBE, tx.incomingAmount))
+
+    def * : ProvenShape[IncomingTransactionDb] =
+      (txIdBE, incomingAmount) <> (fromTuple, toTuple)
+
+    def primaryKey: PrimaryKey =
+      primaryKey("pk_tx", sourceColumns = txIdBE)
+
+    def fk_underlying_tx = {
+      val txTable = TransactionDAO().table
+      foreignKey("fk_underlying_tx",
+        sourceColumns = txIdBE,
+        targetTableQuery = txTable)(_.txIdBE)
+    }
+  }
 }
