+++ wallet/src/main/scala/org/bitcoins/wallet/models/OutgoingTransactionTable.scala
@@ -4,8 +4,6 @@ import org.bitcoins.core.crypto._
 import org.bitcoins.core.currency._
 import org.bitcoins.core.protocol.transaction._
 import org.bitcoins.core.wallet.fee.SatoshisPerByte
-import slick.jdbc.SQLiteProfile.api._
-import slick.lifted.{PrimaryKey, ProvenShape}
 
 /**
   * Represents a relevant transaction for the wallet that we should be keeping track of
@@ -51,66 +49,4 @@ object OutgoingTransactionDb {
       SatoshisPerByte.fromLong(feeRate.toLong)
     )
   }
-}
-
-class OutgoingTransactionTable(tag: Tag)
-    extends Table[OutgoingTransactionDb](tag, "wallet_outgoing_txs")
-    with TxTable[OutgoingTransactionDb] {
-
-  import org.bitcoins.db.DbCommonsColumnMappers._
-
-  def txIdBE: Rep[DoubleSha256DigestBE] = column("txIdBE", O.Unique)
-
-  def inputAmount: Rep[CurrencyUnit] = column("inputAmount")
-
-  def sentAmount: Rep[CurrencyUnit] = column("sentAmount")
-
-  def actualFee: Rep[CurrencyUnit] = column("actualFee")
-
-  def expectedFee: Rep[CurrencyUnit] = column("expectedFee")
-
-  def feeRate: Rep[SatoshisPerByte] = column("feeRate")
-
-  private type OutgoingTransactionTuple =
-    (
-        DoubleSha256DigestBE,
-        CurrencyUnit,
-        CurrencyUnit,
-        CurrencyUnit,
-        CurrencyUnit,
-        SatoshisPerByte)
-
-  private val fromTuple: OutgoingTransactionTuple => OutgoingTransactionDb = {
-    case (txId, inputAmount, sentAmount, actualFee, expectedFee, feeRate) =>
-      OutgoingTransactionDb(txId,
-                            inputAmount,
-                            sentAmount,
-                            actualFee,
-                            expectedFee,
-                            feeRate)
-  }
-
-  private val toTuple: OutgoingTransactionDb => Option[
-    OutgoingTransactionTuple] = tx =>
-    Some(
-      (tx.txIdBE,
-       tx.inputAmount,
-       tx.sentAmount,
-       tx.actualFee,
-       tx.expectedFee,
-       tx.feeRate))
-
-  def * : ProvenShape[OutgoingTransactionDb] =
-    (txIdBE, inputAmount, sentAmount, actualFee, expectedFee, feeRate) <> (fromTuple, toTuple)
-
-  def primaryKey: PrimaryKey =
-    primaryKey("pk_tx", sourceColumns = txIdBE)
-
-  def fk_underlying_tx = {
-    val txTable = TableQuery[TransactionTable]
-    foreignKey("fk_underlying_tx",
-               sourceColumns = txIdBE,
-               targetTableQuery = txTable)(_.txIdBE)
-  }
-
 }
