+++ wallet/src/main/scala/org/bitcoins/wallet/models/SpendingInfoTable.scala
@@ -1,32 +1,13 @@
 package org.bitcoins.wallet.models
 
 import org.bitcoins.core.crypto.{DoubleSha256DigestBE, Sign}
-import org.bitcoins.core.currency.CurrencyUnit
-import org.bitcoins.core.hd.{
-  HDPath,
-  LegacyHDPath,
-  NestedSegWitHDPath,
-  SegWitHDPath
-}
-import org.bitcoins.core.protocol.script.{
-  ScriptPubKey,
-  ScriptWitness,
-  WitnessScriptPubKey
-}
-import org.bitcoins.core.protocol.transaction.{
-  TransactionOutPoint,
-  TransactionOutput
-}
+import org.bitcoins.core.hd.{HDPath, LegacyHDPath, NestedSegWitHDPath, SegWitHDPath}
+import org.bitcoins.core.protocol.script.{ScriptPubKey, ScriptWitness}
+import org.bitcoins.core.protocol.transaction.{TransactionOutPoint, TransactionOutput}
 import org.bitcoins.core.script.crypto.HashType
-import org.bitcoins.core.wallet.utxo.{
-  BitcoinUTXOSpendingInfoFull,
-  ConditionalPath,
-  TxoState
-}
-import org.bitcoins.db.{DbRowAutoInc, TableAutoInc}
+import org.bitcoins.core.wallet.utxo.{BitcoinUTXOSpendingInfoFull, ConditionalPath, TxoState}
+import org.bitcoins.db.DbRowAutoInc
 import org.bitcoins.keymanager.bip39.BIP39KeyManager
-import slick.jdbc.SQLiteProfile.api._
-import slick.lifted.ProvenShape
 
 /**
   * DB representation of a native V0
@@ -200,165 +181,4 @@ sealed trait SpendingInfoDb extends DbRowAutoInc[SpendingInfoDb] {
 
 }
 
-/**
-  * This table stores the necessary information to spend
-  * a transaction output (TXO) at a later point in time. It
-  * also stores how many confirmations it has, whether
-  * or not it is spent (i.e. if it is a UTXO or not) and the
-  * TXID of the transaction that created this output.
-  */
-case class SpendingInfoTable(tag: Tag)
-    extends TableAutoInc[SpendingInfoDb](tag, "txo_spending_info") {
-  import org.bitcoins.db.DbCommonsColumnMappers._
 
-  def outPoint: Rep[TransactionOutPoint] =
-    column("tx_outpoint")
-
-  def txid: Rep[DoubleSha256DigestBE] = column("txid")
-
-  def state: Rep[TxoState] = column("txo_state")
-
-  def scriptPubKey: Rep[ScriptPubKey] = column("script_pub_key")
-
-  def value: Rep[CurrencyUnit] = column("value")
-
-  def privKeyPath: Rep[HDPath] = column("hd_privkey_path")
-
-  def redeemScriptOpt: Rep[Option[ScriptPubKey]] =
-    column("redeem_script")
-
-  def scriptWitnessOpt: Rep[Option[ScriptWitness]] = column("script_witness")
-
-  def blockHash: Rep[Option[DoubleSha256DigestBE]] = column("block_hash")
-
-  /** All UTXOs must have a SPK in the wallet that gets spent to */
-  def fk_scriptPubKey = {
-    val addressTable = TableQuery[AddressTable]
-    foreignKey("fk_scriptPubKey",
-               sourceColumns = scriptPubKey,
-               targetTableQuery = addressTable)(_.scriptPubKey)
-  }
-
-  /** All UTXOs must have a corresponding transaction in the wallet */
-  def fk_incoming_txId = {
-    val txTable = TableQuery[IncomingTransactionTable]
-    foreignKey("fk_incoming_txId",
-               sourceColumns = txid,
-               targetTableQuery = txTable)(_.txIdBE)
-  }
-
-  private type UTXOTuple = (
-      Option[Long], // ID
-      TransactionOutPoint,
-      ScriptPubKey, // output SPK
-      CurrencyUnit, // output value
-      HDPath,
-      Option[ScriptPubKey], // ReedemScript
-      Option[ScriptWitness],
-      TxoState, // state
-      DoubleSha256DigestBE, // TXID
-      Option[DoubleSha256DigestBE] // block hash
-  )
-
-  private val fromTuple: UTXOTuple => SpendingInfoDb = {
-    case (id,
-          outpoint,
-          spk,
-          value,
-          path: SegWitHDPath,
-          None, // ReedemScript
-          Some(scriptWitness),
-          state,
-          txid,
-          blockHash) =>
-      SegwitV0SpendingInfo(
-        outPoint = outpoint,
-        output = TransactionOutput(value, spk),
-        privKeyPath = path,
-        scriptWitness = scriptWitness,
-        id = id,
-        state = state,
-        txid = txid,
-        blockHash = blockHash
-      )
-
-    case (id,
-          outpoint,
-          spk,
-          value,
-          path: LegacyHDPath,
-          None, // RedeemScript
-          None, // ScriptWitness
-          state,
-          txid,
-          blockHash) =>
-      LegacySpendingInfo(outPoint = outpoint,
-                         output = TransactionOutput(value, spk),
-                         privKeyPath = path,
-                         id = id,
-                         state = state,
-                         txid = txid,
-                         blockHash = blockHash)
-
-    case (id,
-          outpoint,
-          spk,
-          value,
-          path: NestedSegWitHDPath,
-          Some(redeemScript), // RedeemScript
-          Some(scriptWitness), // ScriptWitness
-          state,
-          txid,
-          blockHash)
-        if WitnessScriptPubKey.isWitnessScriptPubKey(redeemScript.asm) =>
-      NestedSegwitV0SpendingInfo(outpoint,
-                                 TransactionOutput(value, spk),
-                                 path,
-                                 redeemScript,
-                                 scriptWitness,
-                                 txid,
-                                 state,
-                                 blockHash,
-                                 id)
-
-    case (id,
-          outpoint,
-          spk,
-          value,
-          path,
-          spkOpt,
-          swOpt,
-          spent,
-          txid,
-          blockHash) =>
-      throw new IllegalArgumentException(
-        "Could not construct UtxoSpendingInfoDb from bad tuple:"
-          + s" ($id, $outpoint, $spk, $value, $path, $spkOpt, $swOpt, $spent, $txid, $blockHash).")
-  }
-
-  private val toTuple: SpendingInfoDb => Option[UTXOTuple] =
-    utxo =>
-      Some(
-        (utxo.id,
-         utxo.outPoint,
-         utxo.output.scriptPubKey,
-         utxo.output.value,
-         utxo.privKeyPath,
-         utxo.redeemScriptOpt,
-         utxo.scriptWitnessOpt,
-         utxo.state,
-         utxo.txid,
-         utxo.blockHash))
-
-  def * : ProvenShape[SpendingInfoDb] =
-    (id.?,
-     outPoint,
-     scriptPubKey,
-     value,
-     privKeyPath,
-     redeemScriptOpt,
-     scriptWitnessOpt,
-     state,
-     txid,
-     blockHash) <> (fromTuple, toTuple)
-}
